Added more infos
